fix: service-worker not among puppeteer targets
There was an intermittent issue where the extension's background script (the service-worker) wouldn't show up among the targets reported by puppeteer.

After implementing this with browser.waitForTarget(), the issue has gone away. This also allowed me to remove a couple of functions, so it's a cleaner solution.

And now we no longer need the maxAttemptsToFindExtension argument, so I'm taking that out.